extract common reusable display components
